refactor: rename bindings.ts to decorators.ts
